refactor: profile, persona, likes, following 도메인에 controller, service 폴더링 적용
