+++ src/following/controller/following.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
 import { ApiCreatedResponse, ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
-import baseResponse from '../common/utils/baseResponseStatus';
-import { sucResponse } from '../common/utils/response';
-import { PostFollowRequestDTO } from './dto/post-follow.dto';
-import { FollowingService } from './following.service';
+import baseResponse from '../../common/utils/baseResponseStatus';
+import { sucResponse } from '../../common/utils/response';
+import { PostFollowRequestDTO } from '../dto/post-follow.dto';
+import { FollowingService } from '../service/following.service';
 
 @Controller('follow')
 @ApiTags('Feed API')
