+++ src/likes/service/likes.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
-import { Likes } from '../common/entities/Likes';
-import baseResponse from '../common/utils/baseResponseStatus';
-import { sucResponse } from '../common/utils/response';
-import { PostLikeRequestDTO } from './dto/post-like.dto';
-import { LikesRepository } from './likes.repository';
+import { Likes } from '../../common/entities/Likes';
+import baseResponse from '../../common/utils/baseResponseStatus';
+import { sucResponse } from '../../common/utils/response';
+import { PostLikeRequestDTO } from '../dto/post-like.dto';
+import { LikesRepository } from '../likes.repository';
 
 @Injectable()
 export class LikesService {
