+++ src/persona/persona.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Persona } from "../common/entities/Persona";
-import { CreatePersonaDto } from "./createPersona.dto";
+import { CreatePersonaDto } from "./dto/createPersona.dto";
 import { PersonaModel } from "./persona.model";
 
 @Injectable()
