+++ src/profiles/profiles.module.ts
@@ -5,9 +5,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { PersonaRepository } from '../persona/persona.repository';
 import { Persona } from '../common/entities/Persona';
 import { Profiles } from '../common/entities/Profiles';
-import { ProfilesController } from './profiles.controller';
+import { ProfilesController } from './controller/profiles.controller';
 import { ProfilesRepository } from './profiles.repository';
-import { ProfilesService } from './profiles.service';
+import { ProfilesService } from './service/profiles.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Profiles, Persona])],
